Relations/UpdatingTodoService

+++ src/main/java/com/example/todo/service/TodoService.java
@@ -1,10 +1,13 @@
 package com.example.todo.service;
 
+import com.example.todo.exception.NotAuthorizedException;
 import com.example.todo.exception.NotFoundException;
+import com.example.todo.model.DB.AppUser;
 import com.example.todo.model.DB.Todo;
 import com.example.todo.repo.TodoRepo;
+import com.example.todo.repo.UserRepo;
+import com.example.todo.utils.Helpers;
 import lombok.AllArgsConstructor;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,26 +16,38 @@ import java.util.List;
 @AllArgsConstructor
 public class TodoService {
     private TodoRepo todoRepo;
+    private UserRepo userRepo;
 
 
     public Todo CreateTodo(Todo t){
+        AppUser user = userRepo.findById(Helpers.getAuthedID()).get();
+        t.setOwner(user);
         return todoRepo.save(t);
     }
 
     public List<Todo> getTodos(){
-        return todoRepo.findAll();
+        AppUser user = userRepo.findById(Helpers.getAuthedID()).get();
+        return user.getTodos();
     }
 
     public Todo ToggleTodoStatus(long id){
-        System.out.println(SecurityContextHolder.getContext().getAuthentication().getPrincipal());
         Todo todo = todoRepo.findById(id).orElseThrow(() -> new NotFoundException(id));
-      todo.setCompleted(!todo.isCompleted());
-      todoRepo.save(todo);
-      return todo;
+
+        if(todo.getOwner().getId() != Helpers.getAuthedID()){
+            throw new NotAuthorizedException();
+        }
+
+        todo.setCompleted(!todo.isCompleted());
+        todoRepo.save(todo);
+        return todo;
     }
 
     public void DeleteTodo(long id){
-        todoRepo.findById(id).orElseThrow(() -> new NotFoundException(id));
+        Todo todo = todoRepo.findById(id).orElseThrow(() -> new NotFoundException(id));
+
+        if(todo.getOwner().getId() != Helpers.getAuthedID()){
+            throw new NotAuthorizedException();
+        }
         todoRepo.deleteById(id);
     }
